+++ Services/EventService/src/main/java/io/App/EventService/EventComponent/Community.java
@@ -1,17 +1,23 @@
 package io.App.EventService.EventComponent;
 
+
 public class Community {
 
 	private int id;
 	private String name;
-
-	public Community(int id, String name) {
-		this.id = id;
-		this.name = name;
-	}
-
+	private String description;
+	private int ownerUserId;
+	
 	public Community() {
 		// For REST only
+	}
+
+	public Community(int id, String name, String description,
+			int ownerUserId) {
+		this.id = id;
+		this.name = name;
+		this.description = description;
+		this.ownerUserId = ownerUserId;
 	}
 
 	public int getId() {
@@ -22,12 +28,28 @@ public class Community {
 		return name;
 	}
 
+	public String getDescription() {
+		return description;
+	}
+
+	public int getCommunityOwner() {
+		return ownerUserId;
+	}
+
 	public void setId(int id) {
 		this.id = id;
 	}
 
 	public void setName(String name) {
 		this.name = name;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+
+	public void setCommunityOwner(int ownerUserId) {
+		this.ownerUserId = ownerUserId;
 	}
 
 }