+++ Services/EventService/src/main/java/io/App/EventService/EventComponent/Role.java
@@ -1,5 +1,0 @@
-package io.App.EventService.EventComponent;
-
-public enum Role {
-	ADMIN, EDITOR, USER
-}