+++ Services/EventService/src/main/java/io/App/EventService/EventComponent/UserAuthorizationCheck.java
@@ -1,15 +1,15 @@
-package io.App.CommunityService.communityComponent;
+package io.App.EventService.EventComponent;
 
 import java.util.List;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import io.App.CommunityService.databaseConnection.OperationsDatabaseConnection;
-import io.App.CommunityService.databaseConnection.UserDatabaseConnection;
-import io.App.CommunityService.exceptions.InternalAppException;
-import io.App.CommunityService.exceptions.NonExistantOperationException;
-import io.App.CommunityService.exceptions.UserDoesNotExistException;
-import io.App.CommunityService.exceptions.UserNotAuthorizedException;
+import io.App.EventService.databaseConnection.OperationsDatabaseConnection;
+import io.App.EventService.databaseConnection.UserDatabaseConnection;
+import io.App.EventService.exceptions.InternalAppException;
+import io.App.EventService.exceptions.NonExistantOperationException;
+import io.App.EventService.exceptions.UserDoesNotExistException;
+import io.App.EventService.exceptions.UserNotAuthorizedException;
 
 @SpringBootApplication
 public class UserAuthorizationCheck {