+++ Services/EventService/src/main/java/io/App/EventService/controller/EventCatalogController.java
@@ -19,14 +19,16 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import io.App.EventService.EventComponent.Event;
 import io.App.EventService.EventComponent.EventCatalog;
-import io.App.EventService.EventComponent.Role;
-import io.App.EventService.EventComponent.UserCredentialCheck;
+import io.App.EventService.EventComponent.EventMapper;
+import io.App.EventService.EventComponent.UserAuthorizationCheck;
 import io.App.EventService.dto.EventDTO;
 import io.App.EventService.dto.EventListWrapper;
 import io.App.EventService.dto.Pair;
 import io.App.EventService.exceptions.EventAlreadyExistsException;
 import io.App.EventService.exceptions.InternalAppException;
+import io.App.EventService.exceptions.NonExistantOperationException;
 import io.App.EventService.exceptions.UserDoesNotExistException;
+import io.App.EventService.exceptions.UserNotAuthorizedException;
 
 @RestController
 @RequestMapping("/eventApi")
@@ -35,13 +37,25 @@ public class EventCatalogController {
 	@Autowired
 	private EventCatalog eC;
 	@Autowired
-	private UserCredentialCheck uCC;
+	private UserAuthorizationCheck uAC;
 
 	private static final String INTERNAL_APP_ERROR_MESSAGE = "Internal Application Error";
 
 	@GetMapping("/events")
-	public EventListWrapper eventList() {
-		return this.eC.getAllEvents();
+	public ResponseEntity<Pair<String, EventListWrapper>> eventList() {
+		EventListWrapper eLW = null;
+		try {
+			eLW = new EventListWrapper(EventMapper
+					.eventListToEventDTOListMapper(eC.getAllEvents()));
+		} catch (InternalAppException e) {
+			System.err.println(e.getMessage());
+			return new ResponseEntity<>(new Pair<>(e.getMessage(), null),
+					HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+
+		System.out.println("Successfull events request");
+		return new ResponseEntity<>(
+				new Pair<>("Successfull events request", eLW), HttpStatus.OK);
 	}
 
 	@GetMapping("getEventsFromCommunity/{cID}")
@@ -49,7 +63,9 @@ public class EventCatalogController {
 			@PathVariable("cID") int cID) {
 		EventListWrapper eLW = null;
 		try {
-			eLW = this.eC.getEventsFromCommunity(cID);
+			eLW = new EventListWrapper(
+					EventMapper.eventListToEventDTOListMapper(
+							eC.getEventsFromCommunity(cID)));
 		} catch (InternalAppException e) {
 			System.err.println(e.getMessage());
 			return new ResponseEntity<>(new Pair<>(e.getMessage(), null),
@@ -70,13 +86,13 @@ public class EventCatalogController {
 
 		try {
 			eDTO = objectMapper.readValue(eventJSON, EventDTO.class);
-			Role userRole = this.uCC.getUserRole(eDTO.getOwnerUserName());
-			if (userRole.equals(Role.EDITOR) || userRole.equals(Role.ADMIN)) {
-				this.eC.registerNewEvent(
-						new Event(eDTO.getName(), eDTO.getStart(),
-								eDTO.getEnd(), Integer.parseInt(eDTO.getcID()),
-								eDTO.getcName(), eDTO.getOwnerUserName()));
-			}
+
+			// check user authorization to perform a creation
+			uAC.checkCreateAuthorization(eDTO.getCurrentUser());
+
+			this.eC.registerNewEvent(new Event(eDTO.getName(), eDTO.getStart(),
+					eDTO.getEnd(), eDTO.getCommunity(), eDTO.getEventOwner()));
+
 		} catch (JsonParseException | JsonMappingException
 				| NumberFormatException e) {
 			System.err.println(e.getMessage());
@@ -86,8 +102,9 @@ public class EventCatalogController {
 			System.err.println(e.getMessage());
 			return new ResponseEntity<>(INTERNAL_APP_ERROR_MESSAGE,
 					HttpStatus.INTERNAL_SERVER_ERROR);
-		} catch (UserDoesNotExistException | EventAlreadyExistsException
-				| InternalAppException e) {
+		} catch (EventAlreadyExistsException | InternalAppException
+				| UserDoesNotExistException | UserNotAuthorizedException
+				| NonExistantOperationException e) {
 			System.err.println(e.getMessage());
 			return new ResponseEntity<>(e.getMessage(), HttpStatus.OK);
 		}
@@ -104,10 +121,14 @@ public class EventCatalogController {
 
 		try {
 			eDTO = objectMapper.readValue(eventJSON, EventDTO.class);
-			this.eC.deleteEvent(new Event(Integer.parseInt(eDTO.getId()),
-					eDTO.getName(), eDTO.getStart(), eDTO.getEnd(),
-					Integer.parseInt(eDTO.getcID()), eDTO.getcName(),
-					eDTO.getOwnerUserName()));
+
+			// check user authorization to perform a creation
+			uAC.checkDeleteAuthorization(eDTO.getEventOwner(),
+					eDTO.getCurrentUser());
+
+			this.eC.deleteEvent(new Event(eDTO.getId(), eDTO.getName(),
+					eDTO.getStart(), eDTO.getEnd(), eDTO.getCommunity(),
+					eDTO.getEventOwner()));
 		} catch (JsonParseException | JsonMappingException
 				| NumberFormatException e) {
 			System.err.println(e.getMessage());
@@ -117,13 +138,36 @@ public class EventCatalogController {
 			System.err.println(e.getMessage());
 			return new ResponseEntity<>(INTERNAL_APP_ERROR_MESSAGE,
 					HttpStatus.INTERNAL_SERVER_ERROR);
-		} catch (InternalAppException e) {
+		} catch (InternalAppException | UserDoesNotExistException
+				| UserNotAuthorizedException
+				| NonExistantOperationException e) {
 			System.err.println(e.getMessage());
 			return new ResponseEntity<>(e.getMessage(), HttpStatus.OK);
 		}
 
 		System.out.println("Successfully deleted the Event");
 		return new ResponseEntity<>("Successfully deleted the Event",
+				HttpStatus.OK);
+	}
+
+	@GetMapping(path = "/userCreatedEvents/{userName}")
+	public ResponseEntity<Pair<String, EventListWrapper>> userCreatedEvents(
+			@PathVariable("userName") String userName) {
+		EventListWrapper eLW = null;
+		System.out.println(userName);
+		try {
+			eLW = new EventListWrapper(
+					EventMapper.eventListToEventDTOListMapper(
+							eC.eventsCreatedByUser(userName)));
+		} catch (InternalAppException e) {
+			System.err.println(e.getMessage());
+			return new ResponseEntity<>(new Pair<>(e.getMessage(), null),
+					HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+
+		System.out.println("Successfull created events request");
+		return new ResponseEntity<>(
+				new Pair<>("Successfull created events request", eLW),
 				HttpStatus.OK);
 	}
 }