+++ Services/EventService/src/main/java/io/App/EventService/exceptions/NonExistantOperationException.java
@@ -1,0 +1,10 @@
+package io.App.EventService.exceptions;
+
+public class NonExistantOperationException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+
+	public NonExistantOperationException() {
+		super("The operation you are trying to get does not exist");
+	}
+}